# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Products CD",
    "version": "1.0.0",
    "description": "API de produtos fictícios para treinar a criação de uma API REST com Node.js e Express.",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "xml": {
          "name": "Product"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Documentação do Swagger",
    "url": "http://swagger.io"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local (HTTP)"
    }
  ],
  "tags": [
    {
      "name": "Products",
      "description": "Rotas de produtos"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "summary": "Listar produtos",
        "tags": ["Products"],
        "description": "Obter lista de produtos",
        "responses": {
          "400": {
            "description": "Nenhum produto encontrado"
          },
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Criar produto",
        "tags": ["Products"],
        "description": "Criar um novo produto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                },
                "required": ["name", "price"],
                "example": {
                  "name": "Televisão",
                  "price": 1499.99
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Este produto já existe"
          },
          "401": {
            "description": "Não autorizado"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Obter produto",
        "tags": ["Products"],
        "description": "Obter um produto específico pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Produto não encontrado"
          },
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Atualizar produto",
        "tags": ["Products"],
        "description": "Atualizar um produto específico pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                },
                "required": ["name", "price"],
                "example": {
                  "name": "Computador",
                  "price": 3490
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Produto não encontrado"
          },
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "summary": "Deletar produto",
        "tags": ["Products"],
        "description": "Excluir um produto específico pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Produto não encontrado"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}